refactor: move any code directly interacting with `db` into db-connection.js
